% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_key_join}
\alias{df_key_join}
\title{Example Joint Key 1999-2018}
\format{
\subsection{\code{df_key_join}}{

A data frame with 265'326 rows and 7 columns:
\describe{
\item{geoID}{Numeric: BSU Identifier (Not Harmonized)}
\item{Gcluster_id}{Numeric: Harmonized BSU identifier}
\item{Gname}{Char: Name of BSU-cluster}
\item{mun_id}{Numeric: Municipality identifier (Not Harmonized)}
\item{Ccluster_id}{Numeric: Harmonized Municipality identifier}
\item{Cname}{Char: Name of Municipality-cluster}
...
}
}
}
\source{
\url{https://laurincurschellas.github.io/NGeoTrack/reference/EtE_changes.html}
}
\usage{
df_key_join
}
\description{
A data set created the functions provided in this package.
}
\keyword{datasets}
