% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_change.R
\name{gather_change}
\alias{gather_change}
\title{Create Panel of Changes to Administrative Identifiers}
\usage{
gather_change(type = "kommune", from = startYear , to = endYear)
}
\arguments{
\item{type}{A string: \code{"kommune"} or \code{"grunnkrets"}}

\item{from}{An integer: Start year of panel}

\item{to}{An integer: End year of panel}
}
\value{
A 'data.frame' object, with five columns \code{from}, \code{oldName}, \code{to}, \code{newName}, \code{year}
\itemize{
\item \code{from} - originating numeric identifier
\item \code{oldName} - originating name of adm. unit
\item \code{to} - new numeric identifier
\item \code{newName} - new name of adm. unit
\item \code{year} - year, by which the change is enforced (01. January \code{year})
}
}
\description{
Returns a data frame with the all officially reported changes to administrative identifiers,
for the type and time period selected by the user.

The function calls the API of Statistics Norway (SSB) and returns the pre-processed information.
}
\examples{
expl <- gather_change(type = "kommune", from = 1999, to = 2018)

}
