% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_mun_2021}
\alias{shp_mun_2021}
\title{Shape File Municipalities of Norway (2021)}
\format{
\subsection{\code{shp_mun_2021}}{

A data frame with 363 rows and 3 columns:
\describe{
\item{kommune}{Municipality Identifier (Not Harmonized)}
\item{year}{Year}
\item{geometry}{A sfc_Polygon, spacial definition of municipalities}
...
}
}
}
\source{
\url{https://kartkatalog.geonorge.no/}
}
\usage{
shp_mun_2021
}
\description{
A data series by the Norwegian mapping authority, released on 'GEONORGE'.
}
\keyword{datasets}
