% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Status_quo_type.R
\name{status_quo}
\alias{status_quo}
\title{Create Panel of Valid Administrative Identifiers}
\usage{
status_quo(type = "kommune", from = startYear , to = endYear)
}
\arguments{
\item{type}{A string: \code{"kommune"} or \code{"grunnkrets"}}

\item{from}{An integer: Start year of panel}

\item{to}{An integer: End year of panel}
}
\value{
A 'data.frame' object, with four columns \code{geoID}, \code{name}, \code{from}, \code{to}
\itemize{
\item \code{geoID} - current numeric identifier
\item \code{name} - current name of adm. unit
\item \code{from} - startYear of validity (always year on year)
\item \code{to} - endYear
}
}
\description{
Returns a data frame with the full panel of each years' valid administrative identifiers
for the chosen unit-type and time-period.

The function calls the API of Statistics Norway (SSB) and returns the pre-processed information.
}
\examples{
expl <- status_quo(type = "kommune", from = 1999, to = 2018)

}
